move UserInfo and Demographics to Profile
